Add files via upload
A library management system in SQL is a database system designed to efficiently manage the operations of a library. It typically includes tables for storing information about books, library members, transactions (such as borrowing and returning books), and other relevant data. The system allows librarians to easily search for and track books, manage member accounts, generate reports, and perform other administrative tasks to ensure the smooth functioning of the library. SQL queries are used to retrieve, update, and manipulate data within the system.